Stop sending master credentials to slaves
Sharing master credentials to slaves leads to security issues. A better way of allowing the slaves to download their agents is through a secret:
```
/var/lib/jenkins/agent.jar -jnlpUrl" ${JENKINS_URL}"/computer/"${JENKINS_SLAVEHOST}"/slave-agent.jnlp -secret ${JENKINS_SLAVEHOST_SECRET}"  -workDir "/var/lib/jenkins"
```
Also, I can't find this password being used anywhere by the jenkins-slave charm. So removing it now would cause no harm.

There's an open PR on the jenkins-charm to provided that secret through relations and I plan to work on the jenkins-slave charm next.